restrict pages and add style

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
@@ -8,6 +10,9 @@ from .forms import RoomForm
 
 
 def login_page(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -16,6 +21,7 @@ def login_page(request):
             user = User.objects.get(username=username)
         except User.DoesNotExist:
             messages.error(request, 'User does not exist.')
+
         user = authenticate(request, username=username, password=password)
 
         if user is not None:
@@ -35,11 +41,13 @@ def logout_user(request):
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') is not None else ''
+
     rooms = Room.objects.filter(
         Q(topic__name__icontains=q) |
         Q(name__icontains=q) |
         Q(description__icontains=q)
     )
+
     topics = Topic.objects.all()
     room_count = rooms.count()
     context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
@@ -52,8 +60,10 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
+
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid():
@@ -64,9 +74,13 @@ def create_room(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -78,9 +92,15 @@ def update_room(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
+
     return render(request, 'base/delete.html', {'obj': room})
